Add auth-sources variable customization info
co-authored-by: bryce-carson bcars268@mtroyal.ca

This commit adds a sub-section to the ~README.org~ describing the configuration
of the auth-sources variable, which enables Emacs (and Magit) to access Login or
session keyrings managed by a Linux desktop environment.

This prevents the user from needing to write a ~.gitconfig~ or a ~.netrc~ for
storing Git forge authentication tokens hazardously, or with PGP (GPG?).

---

For what it is worth, I still haven't managed to get a push to work to my fork after this. However, I *can* now run `ssh -T git@github.com` within `eshell` successfully. Why I am _prompted_ for username and password is beyond me. Perhaps it is because I'm a git and haven't unlocked the keyring again by relogging in GNOME. :shrug: