Initial commit for Experimental Webkit Banner
This is working, but the URL text needs to be visible for the (with the code as
it is now) for the widget to display properly. Rendering works fine. The speed
is fine.

What doesn't work is that the URL isn't opened. The widget remains empty.
However, using the inspector HTML from some other webpage can be inserted, and
rendering works fine.